refactor(provider): split provider tests into separate files
Split monolithic provider_test.go into per-provider test files:
context_test.go, model_test.go, claude_test.go, cost_test.go,
session_test.go, pwd_test.go.
